Refactored fileops path
